requirements.txt
PyQt5~=5.15.7
numpy~=1.23.4
Pillow~=9.2.0